Finished adding code to induce_v so that it will return velocity curvature information.  Also, tested the code with axisymmetric initial elements, and it appears that the rotations back into standard coordinates work properly.  The subroutines in stream-fn.c are trusted because they come from a standalone program that was certified separately.
Added curvature information to the blobparms data structure.  These
fields are reset in dpos_vel and dpos_vel_fast with all the other
velocity information.

To be done:

(1) Curvature corrections in dpos_vel_gen?  Perhaps this is needed
when merging elements.

(2) When dx==0 and dy==0, there is special code in MP_Direct.  Perhaps
this type of checking belong in induced_v.

(3) Of course, after the velocity field is computed, it needs to be
corrected using the curvature information.