Forgot to add into previous entry
(4) Include curvature terms in fast multipole summation.

Done.  Just finished adding this part to mpsum.  It needs to be tested
still.

Also, removed stream-fn.c calls from induced_vel and induced_veldev so
that the compile would be clean.  Eventually, it is likely that both
these subroutines will go away.

Modifications to global.h, mpcoeffs.c are just to fix the zip code in
the license.