Some major changes here.
The stepper subroutines have been modified to allow one to use
different solvers should the problem become stiff.  There are two
numerical parameters of interest.

a^2 - 1/a^2 < stifftol means the problem is stiff.
a^2 - 1/a^2 < axisymmtol means the blob is essentially axisymmetric so
the orientation should be slaved to the local flow field.

Some bugs have been removed and things have been streamlined a bit.

Started implementing a BDF solver but it does not work yet.